functions/gcs_to_drive/requirements.txt
google-cloud-storage
google-api-python-client
google-auth
